[Backport] Change default Win10 SDK from 10240 to 10586
This is required for Visual Studio 2013 builds to work; version 10240 of the
Win10 SDK are no longer distributed by Microsoft, 10586 is.

Original commit message:
------------------------
When Chromium switched to the Windows 10 SDK it specified the 10240
version but this is missing IVirtualDesktopManager and other types. This
doesn't affect VS 2015 builds because the package the 10586 SDK, but it
does affect builds where DEPOT_TOOLS_WIN_TOOLCHAIN=0 is set.

This change updates the local-toolchain builds to require the 10586 SDK.
Developers are responsible for installing it.

BUG=491424,516306

Review URL: https://codereview.chromium.org/1608013002