Set Seek Cursor correctly on Seek Bar at the beginning of Live Streaming
When a live stream is started at the current live time
HTMLMediaElement::setCurrentTime is not called from V8
bindings (V8HTMLMediaElement) before starting to stream
because at that moment HTML content doesn't know the
current live time of the live stream. This prevents us
from setting seekCursor on the media element and the
current time in webmediaplayer at the beginning in a
correct manner. In this case the current live time is
propagated from V8 in the first occurrence of this function.
To set the incorrect positions correctly in such situations,
requestSeek on MediaPlayerClient to current_timestamp received
here on the first occurrence.