Don't release MediaPlayer on system sleep for live streams
The seekable ranges still need to be updated even when system
goes into sleep state in the case of live streams so that the
player resumes to correct state after wake-up from system sleep.
The fix here involves avoiding releasing/resetting mediaplayer
resources when destructor for the mediaplayer is not called and
it is not a live stream case.